pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "galliard"
version = "0.1.0"
description = "A GNOME client for Music Player Daemon"
authors = [{name = "Jonathan Dieter", email = "jonathan@dieter.ie"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyGObject>=3.40.0",
    "python-mpd2>=3.0.0",
    "pyxdg>=0.27",
    "snapcast>=2.3.7",
]
keywords = ["mpd", "music", "gnome", "gtk4"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Players",
]

[project.urls]
Homepage = "https://github.com/jdieter/galliard"
Repository = "https://github.com/jdieter/galliard"
Issues = "https://github.com/jdieter/galliard/issues"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
]
test = [
    "pytest>=7",
    "pytest-asyncio>=0.21",
]

[tool.setuptools]
script-files = ["bin/galliard"]

[tool.pyright]
reportMissingModuleSource = false

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live_mpd: requires an mpd binary on PATH",
    "gtk: requires Gtk4 + libadwaita + a display",
]
asyncio_mode = "auto"

# pytest-asyncio 0.21 and PyGObject's gi.events both use Python 3.12-ish
# asyncio APIs that 3.14 has deprecated. There's nothing we can do
# in-tree to silence them without pinning newer versions everywhere, so
# scope the ignore rules to those modules only (filterwarnings' last
# field is a regex on the module path).
filterwarnings = [
    "ignore::DeprecationWarning:pytest_asyncio\\.",
    "ignore::DeprecationWarning:gi\\.events",
]
